Fix type in message

+++ src/mesh_gravity.c
@@ -675,7 +675,7 @@ void compute_potential_distributed(struct pm_mesh* mesh, const struct space* s,
     message("Local density field slice has thickness %d.", (int)local_n0);
   if (verbose)
     message("local patch size = %d, local mesh cells = %lld", nr_local_cells,
-            (local_n0 * N * N));
+            (long long)(local_n0 * N * N));
   if (verbose)
     message("Planning the FFT took %.3f %s.",
             clocks_from_ticks(getticks() - tic), clocks_getunit());
